requirements.txt
Flask>=2.0
python-telegram-bot==13.15
google-generativeai>=0.4.0
python-dotenv
